skeleton updates
I was able to create a very basic skeleton for the project that will likely be changed. I was unable to get routing set up completely but there are ways to view the other pages in the localhost. I also added bootstrap into the project to help with formatting and layouts.